requirements.txt
albumentations==1.1.0
colorama==0.4.4
cycler==0.11.0
efficientnet-pytorch==0.7.1
fonttools==4.33.2
imageio==2.17.0
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.4.22
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.64.0
typing_extensions==4.2.0
